Added a return calculator.
The top-level class is a function that takes in a n >= 1 time series and returns a return series. The reason that the number of series can be greater than one is because correcting for dividend payments or calculating an excess return series need more than just the original price series.