Changed necessary inputs to AnalyticFunctions to be per-security, not per-type.
This was necessary because the precise market data and analytic inputs
are going to be dependant on the actual values of the securities, rather than
just the generic types (which is still a key part of looking up in the
function repository).

This has exposed that the entire dependency graph building system is currently
being done *per recalc execution*, which is just horrible. But that's going to
have to be the way it is until we have portfolio refreshes.