Full implementation of input value resolution.
During dependency graph generation, we probe the actual outputs from the function
chosen to provide a concrete mapping from generic (unspecific) input to
concrete (extra-specific) output from input node.
We then provide the value that's expected to the AnalyticFunctionInputs.
The only odd thing is that we have a separate bit for probing the cache
at result population, because we don't have a dependency node there.
I'm not 100% happy with the state of that, but this gets us past this
hurdle.