Reworked the AnalyticFunction stuff to follow new subinterfaces.
AnalyticFunctionDefinition and AnalyticFunctionInvoker have been split
up in to the sub-interfaces for each type of invocation, so that we don't have
to have all the UnsupportedOperationExceptions all over the place.

Also moved the build sub graph stuff into AbstractAnalyticFunction to minimize
the amount of code that goes into the implementations that isn't needed
for the common cases.

As of now, the code actually doesn't even build fully, because the view and
execution work can't handle the difference yet, but I wanted to check this
in as-is.