Snapshots better handle mixed predicate evaluation rules.
Previous state was that analytic value definitions for market data were precisely
the reverse of the rules of market data. Analytic value definitions are:
- Fewer predicates required
- Matches more predicates available.

Live data rules are the opposite, where definitions are:
- More predicates required
- Matches fewer predicates available

For that reason the snapshot provider returns the underlying object, and then
the engine constructs the required value (with definition required) from the object
itself.
This means then that two nodes can each request different analytic value definitions
which are satisfyable by the same market data primitive, and get the same value
with (most importantly), the definition that they require specified.

In doing this I also converted AbstractAnalyticValue to AnalyticValueImpl, which
is a simple concrete subclass which can't handle scaling.