Roll-up changelist based on my reworking of the dependency graph generation work.
This is a whole lot of various changes, culminating in a complete rewrite of
DependencyGraph (the old one is in there until I'm sure the new one works) because
it was just getting far too complicated to work with.
The Viewer doesn't work as of yet still. Have to stop to work on unit testing
at this point.