Only pass the new PositionReference across to calc nodes.
This is a holder for a quantity and a security identity key (which we can do
now that we've got that as a construct) and is better than trying to reuse
Position or something like that to transmit over the wire.

Also moved the jobIdSource up from the DependencyGraphExecutor to the
SingleComputationCycle so that they're unique across all executors in one
cycle.