Added a basic matrix interface.
Has 1D and 2D matrices at the moment. I got tired of converting Double[] to double[] to use the Colt libraries directly in the code, plus this should allow easier switching of matrix libraries. I haven't yet replaced all of the direct mentions of the Colt libraries in the rest of the analytics code, however.