Massive refactoring of VaR code.
This was originally intended to remove the need to calculate the same thing over and over again. However, it became obvious that something more drastic was needed if there wasn't to be huge amounts of repeated code, identical except for one class taking a covariance matrix and sensitivity vector, and the other a P&L timeseries.

There will probably need to be a bit more refinement and testing, but this is how the code will look for VaR.