[IGN-4] Added support for terminal output values on DependencyNode.
The problem with stripping out unnecessary output values was that it didn't leave a place for specifying
that an output value was one that we cared about at the top level of the graph, without any dependent nodes.
This change lets us mark a value as one that we care about no matter what, which is then used.

Re-enables aggregate computations.