[ENG-62] Created new interface for sending calculation jobs.
We started out with an interface like this, and then changed to not using it, but just putting
a FudgeRequestSender into the ViewProcessingContext.

Now, however, to support fast-pass in-JVM mode, we need it back, so I've redone it, with a wrapper
that allows you to use a FudgeRequestSender underneath for compatibility.