Started writing unit tests for PortfolioEvaluationModel.
When I did it I realized that it was quite unwieldy that we needed all the runtime
services (like for job dispatch) just for compilation, so I separated that out
to the new ViewCompilationServices, which is what we need for compilation.

Also took subscriptions to live data out of the PortfolioEvaluationModel and into the View's space.