[ANA-85] Allow clients of Interpolator1D to hold an Interpolator1DModel.
This happens through the new Interpolator1DModelFactory which constructs an optimized instance of Interpolator1DModel based on the particular types of inputs.
Also includes a new InterpolationBoundedValues mini-object (returned by a method on Interpolator1DModel) which is very optimized for Linear interpolation.
Minor testing enhancements and fixes to the code submitted for ANA-86.