Changed name of function in final double price = calculator.getPresentValueLegacyCreditDefaultSwap(now, cds, yieldCurve, hazardRateCurve, _priceType);
This gets the build working again, but potentially at the expense of breaking the logic (this will need to be checked).

This problem arose because of the extensive refactoring of the legacy/standard/vanilla CDS code. Everything in OG Analytics seemed to refactor as expected, but the changes did not apear to be propagated to OG Financial.