[PLAT-3488] NonVersionedRedisHistoricalTimeSeriesSource
The primary and initial implementation of the HistoricalTimeSeries
interface has been completed, and basic test cases done.
Additional methods could be implemented if it turns out that they're
actually invoked at runtime; unfortunately there's so much legacy
cruft in HistoricalTimeSeries that it's hard to tell
what is actually necessary to implement.