[PLAT-3496] Add snapshotting unstructured values when running with hts as input
Previously when running with historical input the depgraph is rewritten to replace the external ids of market values with the hts unique id. This causes the external Id to be unavailable when adding the value to a snapshot and hence the value is missing from the resultant snapshot.

Fix by calling HistoricalTimeSeriesSource.getExternalIdBundle(uniqueId).

This needs the historical time series to be added to the view processor component factory. If it is not specified the behaviour will be as before.