[PLAT-3867] Respect ViewClient's ViewResultMode.
Previous to this change, regardless of whether there were any clients
requesting the results with a ViewResultMode that specifies that
deltas are requested, the previous result will be cached and deltas
will be calculated. For a large result model, this is a significant
amount of heap and delta processing.

This commit ensures that only if any view clients require delta results
will they be calculated or the previous result stored in the heap.