[PLAT-3867] Don't cache results unless necessary.
Until this commit, the ViewProcessMXBeanImpl will cache all results
whether they're going to be used in a processResults() call or not.
This commit makes this a manually controlled process, so that the
view result model is not inadvertantly cached (particularly in
programmatic cases where we're trying to avoid the heap charge
for storing the previous cycle's result model).