[PLAT-2745] deduce which DriverManager method to call based upon whether us er/pass are actually present or not
When using MSSQL, one can connect via microsoft single sign on dlls,
so no user credentials and password are involved in the java code
(just a java.library.path pointing to the DLLs). But our
SQLCatalogCreationStrategy was unconditionally passing user and
password when connecting, even if they were blank or null.
It should call a different method signature if only a url is provided