[PLAT-4895] NonVersionedRedisHistoricalTimeSeriesSource.exist().
In order for the Resolver to operate efficiently, it needs to have a much
faster check on whether the TS exists. Previously it was loading the whole
time series.

This commit changes the behavior so that there's just a redis EXISTS
call made to determine existence of the time series.