[PLAT-4950] Speed up searches by ExternalId in InMemorySecurityMaster.
This is the most common search criteria when using an InMemorySecurityMaster
underneath a MasterSecuritySource for plugging into the engine.
Providing an explicit cache thus has the chance to significantly improve
CPU performance when doing using an InMemorySecurityMaster with large
amounts of securities stored in memory, which is common in several
key use cases.